refactor(sdk): rename reasoning helper module
