+++ openhands-sdk/openhands/sdk/event/llm_convertible/reasoning_utils.py
